+++ src/libexpr/include/nix/expr/meson.build
@@ -31,7 +31,8 @@ headers = [ config_pub_h ] + files(
   'print.hh',
   'repl-exit-status.hh',
   'search-path.hh',
-  'static-string-data.hh',
+  'string-data.hh',
+  'string-data-static.hh',
   'symbol-table.hh',
   'value-to-json.hh',
   'value-to-xml.hh',
