+++ src/libexpr/include/nix/expr/symbol-table.hh
@@ -3,7 +3,7 @@
 
 #include <memory_resource>
 #include "nix/expr/value.hh"
-#include "nix/expr/static-string-data.hh"
+#include "nix/expr/string-data-static.hh"
 #include "nix/util/chunked-vector.hh"
 #include "nix/util/error.hh"
